requirements.txt
matplotlib==3.4.3
numpy==1.22.4
opencv_python==4.5.5.64
pandas==1.3.4
Pillow==9.4.0
PyYAML==6.0
scipy==1.7.2
six==1.16.0
tqdm==4.62.3
pyctcdecode
tensorboard
timm
einops
torch_geometric